move topic idea generation and chat logic to ideas_service for better modularity
